Consolidated datasource code into datasource package
